backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
requests
python-dotenv
